Split widget: panel scroll view
