+++ extension/devtools/app/lib/widgets/panel_scroll_view.dart
@@ -1,0 +1,21 @@
+part of '../main.dart';
+
+class _PanelScrollView extends StatelessWidget {
+  const _PanelScrollView({required this.child});
+
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        return SingleChildScrollView(
+          child: ConstrainedBox(
+            constraints: BoxConstraints(minHeight: constraints.maxHeight),
+            child: child,
+          ),
+        );
+      },
+    );
+  }
+}
